fix: avoided allocation by choosing the param type
This commit is a good example on how choosing a good datatype cleans the code. `match_tokens` do not need to own the list you are passing to it, it only needs to read it. By changing his datatype to `&[T]`, it suddenly become clear that at **all** call sites you were allocating only to call this function. Now all slices are static, and the malloc demon is left alone.