feat: `clap` for parsing command line args
Ok this might be too much for a simple interface like this, but it absolute ease the making of the CLI.

Also, `PathBuf` instead of `String` and `Path` instead of `str`. Paths can contain chars that are not utf-8, depending from the OS. For example on Windows they are utf-16. Path and PathBuf represent native paths, and are fully cross platform. They also provide a nice API to manipulate paths.